Preprocessor hackery completed for now
Tests now correctly don't run unless you're playing and in main screen. Also, they pass now, but statistics panel tests are in early stage and need to be expanded. Need to also add this game version stuff to arrays etc. in the future and handle non-primitive members, but that's for another day.